refactor(articles): extract lifecycle services
